+++ src/Container/EventMiddlewareFactory.php
@@ -11,16 +11,24 @@ namespace Prooph\Psr7Middleware\Container;
 
 use Interop\Config\ConfigurationTrait;
 use Interop\Config\ProvidesDefaultOptions;
-use Interop\Config\RequiresContainerId;
 use Interop\Config\RequiresMandatoryOptions;
 use Interop\Container\ContainerInterface;
 use Prooph\Psr7Middleware\EventMiddleware;
 use Prooph\Psr7Middleware\NoopMetadataGatherer;
 use Prooph\ServiceBus\EventBus;
 
-final class EventMiddlewareFactory implements RequiresContainerId, ProvidesDefaultOptions, RequiresMandatoryOptions
+final class EventMiddlewareFactory extends AbstractMiddlewareFactory
+    implements ProvidesDefaultOptions, RequiresMandatoryOptions
 {
     use ConfigurationTrait;
+
+    /**
+     * @param string $configId
+     */
+    public function __construct($configId = 'event')
+    {
+        parent::__construct($configId);
+    }
 
     /**
      * Create service.
@@ -30,7 +38,7 @@ final class EventMiddlewareFactory implements RequiresContainerId, ProvidesDefau
      */
     public function __invoke(ContainerInterface $container)
     {
-        $options = $this->options($container->get('config'));
+        $options = $this->options($container->get('config'), $this->configId);
 
         if (isset($options['metadata_gatherer'])) {
             $gatherer = $container->get($options['metadata_gatherer']);
@@ -43,30 +51,6 @@ final class EventMiddlewareFactory implements RequiresContainerId, ProvidesDefau
             $container->get($options['message_factory']),
             $gatherer
         );
-    }
-
-    /**
-     * @interitdoc
-     */
-    public function vendorName()
-    {
-        return 'prooph';
-    }
-
-    /**
-     * @interitdoc
-     */
-    public function packageName()
-    {
-        return 'middleware';
-    }
-
-    /**
-     * @interitdoc
-     */
-    public function containerId()
-    {
-        return 'event';
     }
 
     /**