+++ CHANGELOG.md
@@ -2,11 +2,47 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
-## 0.1.0 TBA
+## 0.3.0 TBA
 
 ### Added
 
-* Middleware for command, query, event and message
+* [#9](https://github.com/prooph/psr7-middleware/pull/8) Use interop-config version 1.0
+
+### Deprecated
+
+* Nothing
+
+### Removed
+
+* Nothing
+
+### Fixed
+
+* Nothing
+
+## 0.2.0 (2016-02-09)
+
+### Added
+
+* [#7](https://github.com/prooph/psr7-middleware/pull/7) Introduce a MetadataGatherer concept
+
+### Deprecated
+
+* Nothing
+
+### Removed
+
+* Nothing
+
+### Fixed
+
+* Nothing
+
+## 0.1.0 (2016-02-09)
+
+### Added
+
+* [#1](https://github.com/prooph/psr7-middleware/pull/1) Middleware for command, query, event and message
 
 ### Deprecated